Repurpose some old UriNotifier tests to test DataChangedNotifiers more generically
